:tractor: Reorganize clay a bit and remove repository interface
