+++ pkg/cmds/ls-commands/ls-commands.go
@@ -1,0 +1,3 @@
+package ls_commands
+
+type ListCommandsCommand struct{}
