--- pkg/repositories/repository.go
@@ -1,7 +1,0 @@
-package repositories
-
-import "github.com/go-go-golems/glazed/pkg/cmds"
-
-type Repository interface {
-	CollectCommands(prefix []string, recurse bool) []cmds.Command
-}
